+++ playwright/suites/multiparty-conference-set-9-tests.spec.ts
@@ -1,0 +1,4 @@
+import {test} from '@playwright/test';
+import createMultipartyConferenceSet9Tests from '../tests/multiparty-conference-set-9-test.spec';
+
+test.describe('Multiparty Conference Tests - Set 9', createMultipartyConferenceSet9Tests);
