+++ playwright/test-data.ts
@@ -95,4 +95,17 @@ export const USER_SETS = {
     ENTRY_POINT: env.PW_ENTRY_POINT8,
     TEST_SUITE: 'multiparty-conference-set-8-tests.spec.ts',
   },
+  SET_9: {
+    AGENTS: {
+      AGENT1: {username: 'user33', extension: '1033', agentName: 'User33 Agent33'},
+      AGENT2: {username: 'user34', extension: '1034', agentName: 'User34 Agent34'},
+      AGENT3: {username: 'user35', extension: '1035', agentName: 'User35 Agent35'},
+      AGENT4: {username: 'user36', extension: '1036', agentName: 'User36 Agent36'},
+    },
+    QUEUE_NAME: 'Queue e2e 9',
+    CHAT_URL: `${env.PW_CHAT_URL}-e2e-9.html`,
+    EMAIL_ENTRY_POINT: `${env.PW_SANDBOX}.e2e9@gmail.com`,
+    ENTRY_POINT: env.PW_ENTRY_POINT9,
+    TEST_SUITE: 'multiparty-conference-set-9-tests.spec.ts',
+  },
 };
