+++ playwright/tests/multiparty-conference-set-7-test.spec.ts
@@ -1,21 +1,17 @@
 import {test, expect, Page} from '@playwright/test';
 import {TestManager} from '../test-manager';
 import {verifyCurrentState} from '../Utils/userStateUtils';
-import {acceptIncomingTask, declineIncomingTask, endCallTask, waitForIncomingTask} from '../Utils/incomingTaskUtils';
-import {consultOrTransfer, cancelConsult} from '../Utils/advancedTaskControlUtils';
-import {submitWrapup} from '../Utils/wrapupUtils';
+import {endCallTask} from '../Utils/incomingTaskUtils';
+import {waitForState} from '../Utils/helperUtils';
 import {
   AgentId,
   CONFERENCE_AGENT_IDS,
-  cleanupConferencePageWithTimeout,
   cleanupConferenceState,
   consultAgentAndAcceptCall,
-  consultQueueAndAcceptCall,
   exitConferenceParticipantOrEndTask,
   getConferenceAgentName,
   getConferenceRequiredEnv,
   mergeConsultIntoConference as mergeConsultIntoConferenceUtil,
-  setConferenceAgentsAvailable,
   startBaselineCallOnAgent1 as startBaselineCallOnAgent1Util,
   toggleConferenceLegIfSwitchAvailable,
   transferConsultAndSubmitWrapup,
@@ -25,10 +21,8 @@ import {
   CONFERENCE_ACTION_SETTLE_TIMEOUT,
   CONFERENCE_CUSTOMER_DISCONNECT_TIMEOUT,
   CONFERENCE_RECONNECT_SETTLE_TIMEOUT,
-  CONSULT_NO_ANSWER_TIMEOUT,
-  TASK_TYPES,
+  CONFERENCE_SWITCH_TOGGLE_TIMEOUT,
   USER_STATES,
-  WRAPUP_REASONS,
 } from '../constants';
 
 export default function createMultipartyConferenceSet7Tests() {
@@ -54,8 +48,6 @@ export default function createMultipartyConferenceSet7Tests() {
   };
 
   const getAgentName = (agentId: AgentId): string => getConferenceAgentName(testManager.projectName, agentId);
-  const setConsultTargetAgentsAvailable = async (agentIds: AgentId[]) =>
-    setConferenceAgentsAvailable(getAgentPage, agentIds);
 
   const cleanupConferenceStateForTest = async () => {
     await cleanupConferenceState({
@@ -80,15 +72,6 @@ export default function createMultipartyConferenceSet7Tests() {
       toAgent,
       getAgentPage,
       getAgentName,
-    });
-  };
-
-  const consultQueueAndAcceptIncomingCall = async (fromAgent: AgentId, toAgent: AgentId) => {
-    await consultQueueAndAcceptCall({
-      fromAgent,
-      toAgent,
-      getAgentPage,
-      getRequiredEnv,
     });
   };
 
@@ -228,75 +211,41 @@ export default function createMultipartyConferenceSet7Tests() {
     test.skip('CTS-MPC-08 should validate conference participant limit (>4 agents required)', async () => {});
   });
 
-  test.describe('TRANSFER CONFERENCE SCENARIOS - Set 7', () => {
-    test('CTS-TC-01, CTS-TC-02 and CTS-TC-03 should handle blind transfer to agent, then queue transfer in one call session', async () => {
+  test.describe('Switch Conference - Set 7', () => {
+    test('CTS-SW-02 and CTS-SW-03 should switch legs and merge lobby participants into conference', async () => {
       await startBaselineCallOnAgent1();
-      await setConsultTargetAgentsAvailable([2]);
-
-      await consultOrTransfer(getAgentPage(1), 'agent', 'transfer', getAgentName(2));
-      await acceptIncomingTask(getAgentPage(2), TASK_TYPES.CALL, ACCEPT_TASK_TIMEOUT * 2);
-      await submitWrapup(getAgentPage(1), WRAPUP_REASONS.SALE);
-      await verifyCurrentState(getAgentPage(2), USER_STATES.ENGAGED);
-
-      await setConsultTargetAgentsAvailable([1]);
-      await consultOrTransfer(getAgentPage(2), 'queue', 'transfer', getRequiredEnv('QUEUE_NAME'));
-      await acceptIncomingTask(getAgentPage(1), TASK_TYPES.CALL, ACCEPT_TASK_TIMEOUT * 2);
-      await submitWrapup(getAgentPage(2), WRAPUP_REASONS.SALE);
-
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
-    });
-
-    test('CTS-TC-04 and CTS-TC-05 should run consult accepted and declined flows in one call session', async () => {
-      await startBaselineCallOnAgent1();
-
+      await consultAgentAndAccept(1, 2);
+      await mergeConsultIntoConference(1);
       await consultAgentAndAccept(1, 3);
-      await cancelConsult(getAgentPage(3));
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
 
-      await consultOrTransfer(getAgentPage(1), 'agent', 'consult', getAgentName(3));
-      await waitForIncomingTask(getAgentPage(3), TASK_TYPES.CALL, ACCEPT_TASK_TIMEOUT);
-      await declineIncomingTask(getAgentPage(3), TASK_TYPES.CALL);
+      const page = getAgentPage(1);
+      const switchToMainButton = page.getByTestId('switchToMainCall-consult-btn');
+      const switchToConsultButton = page.getByTestId('call-control:switch-to-consult').first();
+
+      await expect(switchToMainButton).toBeVisible({timeout: ACCEPT_TASK_TIMEOUT});
+      await switchToMainButton.click();
+      await page.waitForTimeout(CONFERENCE_SWITCH_TOGGLE_TIMEOUT);
+
+      await expect(switchToConsultButton).toBeVisible({timeout: ACCEPT_TASK_TIMEOUT});
+      await switchToConsultButton.click();
+      await page.waitForTimeout(CONFERENCE_SWITCH_TOGGLE_TIMEOUT);
+
+      await mergeConsultIntoConference(1);
       await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
+      await verifyCurrentState(getAgentPage(3), USER_STATES.ENGAGED);
     });
 
-    // Split: queue routing won't re-route to an agent who RONA'd in the same session
-    test('CTS-TC-06 should handle not-picked agent consult', async () => {
+    test('CTS-SW-04 should transfer from consult lobby to conference participant', async () => {
       await startBaselineCallOnAgent1();
-      await setConsultTargetAgentsAvailable([2]);
-
-      await consultOrTransfer(getAgentPage(1), 'agent', 'consult', getAgentName(2));
-      await getAgentPage(1).waitForTimeout(CONSULT_NO_ANSWER_TIMEOUT);
-
-      const cancelConsultButton = getAgentPage(1).getByTestId('cancel-consult-btn');
-      const canCancelConsult = await cancelConsultButton.isVisible().catch(() => false);
-      if (canCancelConsult) {
-        await cancelConsult(getAgentPage(1));
-      }
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
-    });
-
-    // Use agent 3 instead of agent 2 — queue routing won't re-route to an
-    // agent who RONA'd (CTS-TC-06) in the same session.
-    test('CTS-TC-07 should handle queue-consult cancel', async () => {
-      await startBaselineCallOnAgent1();
-      await consultQueueAndAcceptIncomingCall(1, 3);
-      await cancelConsult(getAgentPage(1));
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
-    });
-
-    // Use agent 3 instead of agent 2 — agent 2 RONA'd in CTS-TC-06 and may
-    // not reliably receive consults for the remainder of the session.
-    test('CTS-TC-08 should support multi-stage consult transfer between A1 and A3', async () => {
-      await startBaselineCallOnAgent1();
-
+      await consultAgentAndAccept(1, 2);
+      await mergeConsultIntoConference(1);
       await consultAgentAndAccept(1, 3);
       await transferConsultAndWrapup(1);
-      await verifyCurrentState(getAgentPage(3), USER_STATES.ENGAGED);
 
-      await cleanupConferencePageWithTimeout(getAgentPage(1), testManager.callerPage);
-      await consultAgentAndAccept(3, 1);
-      await transferConsultAndWrapup(3);
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
+      await waitForState(getAgentPage(2), USER_STATES.ENGAGED);
+      await waitForState(getAgentPage(3), USER_STATES.ENGAGED);
+      await verifyCurrentState(getAgentPage(2), USER_STATES.ENGAGED);
+      await verifyCurrentState(getAgentPage(3), USER_STATES.ENGAGED);
     });
   });
 }
