+++ playwright/tests/multiparty-conference-set-8-test.spec.ts
@@ -4,7 +4,6 @@ import {verifyCurrentState} from '../Utils/userStateUtils';
 import {acceptIncomingTask, endCallTask} from '../Utils/incomingTaskUtils';
 import {consultOrTransfer, cancelConsult} from '../Utils/advancedTaskControlUtils';
 import {submitWrapup} from '../Utils/wrapupUtils';
-import {waitForState} from '../Utils/helperUtils';
 import {
   AgentId,
   CONFERENCE_AGENT_IDS,
@@ -19,13 +18,7 @@ import {
   startBaselineCallOnAgent1 as startBaselineCallOnAgent1Util,
   transferConsultAndSubmitWrapup,
 } from '../Utils/conferenceUtils';
-import {
-  ACCEPT_TASK_TIMEOUT,
-  CONFERENCE_SWITCH_TOGGLE_TIMEOUT,
-  TASK_TYPES,
-  USER_STATES,
-  WRAPUP_REASONS,
-} from '../constants';
+import {ACCEPT_TASK_TIMEOUT, TASK_TYPES, USER_STATES, WRAPUP_REASONS} from '../constants';
 
 export default function createMultipartyConferenceSet8Tests() {
   let testManager: TestManager;
@@ -183,68 +176,7 @@ export default function createMultipartyConferenceSet8Tests() {
     test.skip('CTS-TC-18 should validate EPDN transfer conference handoff flow', async () => {});
   });
 
-  test.describe('Switch Conference', () => {
-    test.skip('CTS-SW-01 should switch conference via EP_DN handoff', async () => {});
-
-    test('CTS-SW-02 and CTS-SW-03 should switch legs and merge lobby participants into conference', async () => {
-      await startBaselineCallOnAgent1();
-      await consultAgentAndAccept(1, 2);
-      await mergeConsultIntoConference(1);
-      await consultAgentAndAccept(1, 3);
-
-      const page = getAgentPage(1);
-      const switchToMainButton = page.getByTestId('switchToMainCall-consult-btn');
-      const switchToConsultButton = page.getByTestId('call-control:switch-to-consult').first();
-
-      // Switch from consult leg to main call leg
-      await expect(switchToMainButton).toBeVisible({timeout: ACCEPT_TASK_TIMEOUT});
-      await switchToMainButton.click();
-      await page.waitForTimeout(CONFERENCE_SWITCH_TOGGLE_TIMEOUT);
-
-      // Switch back from main call leg to consult leg
-      await expect(switchToConsultButton).toBeVisible({timeout: ACCEPT_TASK_TIMEOUT});
-      await switchToConsultButton.click();
-      await page.waitForTimeout(CONFERENCE_SWITCH_TOGGLE_TIMEOUT);
-
-      await mergeConsultIntoConference(1);
-      await verifyCurrentState(getAgentPage(1), USER_STATES.ENGAGED);
-      await verifyCurrentState(getAgentPage(3), USER_STATES.ENGAGED);
-    });
-
-    test('CTS-SW-04 should transfer from consult lobby to conference participant', async () => {
-      await startBaselineCallOnAgent1();
-      await consultAgentAndAccept(1, 2);
-      await mergeConsultIntoConference(1);
-      await consultAgentAndAccept(1, 3);
-      await transferConsultAndWrapup(1);
-
-      // Transfer handoff settles asynchronously; wait for both participants to reflect engaged state.
-      await waitForState(getAgentPage(2), USER_STATES.ENGAGED);
-      await waitForState(getAgentPage(3), USER_STATES.ENGAGED);
-      await verifyCurrentState(getAgentPage(2), USER_STATES.ENGAGED);
-      await verifyCurrentState(getAgentPage(3), USER_STATES.ENGAGED);
-    });
-
-    test('CTS-SW-05 and CTS-SW-06 should block consult during active consult and restore it after lobby closes', async () => {
-      await startBaselineCallOnAgent1();
-      await consultAgentAndAccept(1, 2);
-      await mergeConsultIntoConference(1);
-
-      await consultAgentAndAccept(1, 3);
-      const agent2Consult = getAgentPage(2).getByTestId('call-control:consult').first();
-      const consultVisible = await agent2Consult.isVisible().catch(() => false);
-
-      if (consultVisible) {
-        await expect(agent2Consult).toBeDisabled();
-      } else {
-        expect(consultVisible).toBeFalsy();
-      }
-
-      await mergeConsultIntoConference(1);
-
-      await expect(agent2Consult).toBeVisible({timeout: ACCEPT_TASK_TIMEOUT});
-    });
-
+  test.describe('Switch Conference - Set 8', () => {
     test('CTS-SW-07 should keep queue-based switch flow stable with consult cancel/accept actions', async () => {
       await startBaselineCallOnAgent1();
       await setConsultTargetAgentsAvailable([2]);
@@ -257,7 +189,5 @@ export default function createMultipartyConferenceSet8Tests() {
       await mergeConsultIntoConference(1);
       await verifyCurrentState(getAgentPage(2), USER_STATES.ENGAGED);
     });
-
-    test.skip('CTS-SW-08 should validate switch conference flows with >4 agents', async () => {});
   });
 }
